integer a long

+++ src/main/java/uniandes/edu/co/proyecto/Controller/OrdenController.java
@@ -41,7 +41,7 @@ public class OrdenController {
     }
     
     @GetMapping("/ordenes/{id}/edit/save")
-    public ResponseEntity<String> ordenAnularGuardar(@PathVariable("id")Integer id, @RequestBody Orden orden){
+    public ResponseEntity<String> ordenAnularGuardar(@PathVariable("id")Long id, @RequestBody Orden orden){
         try{
             ordenRepository.actualizarOrdenAnulada(id, orden.getFechaEstimada(), orden.getFechaLlegada(), orden.getFechaCreacion(), orden.getProveedor().getNIT(), orden.getBodegaDestino().getId());
             return new ResponseEntity<>("Orden actualizada exitosamente", HttpStatus.OK);
